fix: Se cambio el estado de clientes de Int a Long

+++ src/main/java/com/pablosipac/kinalapp/entity/Cliente.java
@@ -19,12 +19,12 @@ public class    Cliente {
     @Column
     private String direccion;
     @Column
-    private int estado;
+    private Long estado;
 
     public Cliente() {
     }
 
-    public Cliente(String DPICliente, String nombreCliente, String apellidoCliente, String direccion, int estado) {
+    public Cliente(String DPICliente, String nombreCliente, String apellidoCliente, String direccion, Long estado) {
         this.DPICliente = DPICliente;
         this.nombreCliente = nombreCliente;
         this.apellidoCliente = apellidoCliente;
@@ -64,11 +64,11 @@ public class    Cliente {
         this.direccion = direccion;
     }
 
-    public int getEstado() {
+    public Long getEstado() {
         return estado;
     }
 
-    public void setEstado(int estado) {
+    public void setEstado(Long estado) {
         this.estado = estado;
     }
 }
